+++ simplecrypto-cxx/pbkdf2.h
@@ -1,8 +1,36 @@
 #ifndef PBKDF2_H
 #define PBKDF2_H
 
-
-//#include "sha2.hpp"
+/**
+ * Copyright (c) 2000-2001 Aaron D. Gifford
+ * Copyright (c) 2013-2014 Pavol Rusnak
+ * Copyright (c) 2020 Waqar Ahmed
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. Neither the name of the copyright holder nor the names of contributors
+ *    may be used to endorse or promote products derived from this software
+ *    without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTOR(S) ``AS IS'' AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED.  IN NO EVENT SHALL THE AUTHOR OR CONTRIBUTOR(S) BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
+ * OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
+ * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
+ * SUCH DAMAGE.
+ */
 #include "sha256.h"
 #include "sha512.h"
 
@@ -19,11 +47,6 @@ typedef struct _PBKDF2_HMAC_SHA512_CTX {
     std::array<uint64_t, SHA512_RAW_BYTES_LENGTH / sizeof(uint64_t)> idig;
     std::array<uint64_t, SHA512_RAW_BYTES_LENGTH / sizeof(uint64_t)> f;
     std::array<uint64_t, SHA512_BLOCK_LENGTH / sizeof(uint64_t)> g;
-    //    std::array<uint64_t, 16> g;
-    //    uint64_t odig[SHA512_RAW_BYTES_LENGTH / sizeof(uint64_t)];
-    //    uint64_t idig[SHA512_RAW_BYTES_LENGTH / sizeof(uint64_t)];
-    //    uint64_t f[SHA512_RAW_BYTES_LENGTH / sizeof(uint64_t)];
-    //    uint64_t g[SHA512_BLOCK_LENGTH / sizeof(uint64_t)];
     char first;
 } PBKDF2_HMAC_SHA512_CTX;
 
@@ -44,6 +67,11 @@ void pbkdf2_hmac_sha512_Init(
 void pbkdf2_hmac_sha512_Update(PBKDF2_HMAC_SHA512_CTX* pctx, uint32_t iterations);
 void pbkdf2_hmac_sha512_Final(PBKDF2_HMAC_SHA512_CTX* pctx, uint8_t* key);
 void pbkdf2_hmac_sha512(
-    const uint8_t* pass, int passlen, const uint8_t* salt, int saltlen, uint32_t iterations, uint8_t* key);
+    const uint8_t* pass,
+    size_t passlen,
+    const uint8_t* salt,
+    size_t saltlen,
+    uint32_t iterations,
+    uint8_t* key);
 
 #endif // PBKDF2_H
